Changed chatroom_id to bigint, removed extra char
- Due to the old method of using user_id as the id for the chatroom, it
has to be changed to a biint to accomodate the large values of user_id
- removed an extra parentheses in QueryManager

+++ SQL/eatsee_db_20230511.sql
@@ -56,7 +56,7 @@ CREATE TABLE gender_ref(
 );
 
 CREATE TABLE chat(
-	chatroom_id  INT    NOT NULL,
+	chatroom_id  BIGINT    NOT NULL,
 	user_id		 BIGINT    NOT NULL UNIQUE,
 	CONSTRAINT fk_chat_user_users
 		FOREIGN KEY(user_id)
